% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CVer.R
\name{CVer}
\alias{CVer}
\title{Calculating % coefficient of variation from mixed model}
\usage{
CVer(model)
}
\arguments{
\item{model}{An object of class \code{merMod}}
}
\description{
Calculating % coefficient of variation from mixed model
}
\keyword{CV}
\keyword{lmer}
