% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HSDer.R
\name{HSDer}
\alias{HSDer}
\title{Calculating Tukey's HSD (Honestly Significant Difference) value from a mixed effects regression model fitted using \code{lme4}}
\usage{
HSDer(model, term, comps = NULL, level = 0.95)
}
\arguments{
\item{model}{An object of class \code{merMod}}

\item{term}{string indicating variable to calculate HSDs for}

\item{comps}{(optional) vector of length 2 indicating which groups to compare}

\item{level}{confidence level for HSD}
}
\description{
Calculating Tukey's HSD (Honestly Significant Difference) value from a mixed effects regression model fitted using \code{lme4}
}
\keyword{HSD}
\keyword{Tukey}
\keyword{lmer}
