% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LSDer.R
\name{LSDer}
\alias{LSDer}
\title{Calculating Fishers LSD value from mixed model}
\usage{
LSDer(model, term, comps = NULL, level = 0.95)
}
\arguments{
\item{model}{An object of class \code{merMod}}

\item{term}{string indicating variable to calculate LSDs for}

\item{comps}{vector of length 2 indicating which groups to compare}

\item{level}{confidence level for LSD}
}
\description{
Calculating Fishers LSD value from mixed model
}
\keyword{LSD}
\keyword{lmer}
